refactor: Move MD003 configuration to rule module for better locality
This refactoring demonstrates a new architecture pattern for rule-specific configuration:

## Changes Made
- **Moved MD003 config types** from `config/mod.rs` to `rules/md003.rs`
  - HeadingStyle enum with all variants (consistent, atx, setext, etc.)
  - MD003HeadingStyleTable struct with serde derives
- **Added re-export mechanism** in config/mod.rs for backward compatibility
- **Updated imports** in MD003 rule and tests to use local config types
- **Consolidated configuration management** while removing the separate quickmark_config crate

## Architecture Benefits
- **Better locality of reference**: Rule configuration co-located with implementation
- **Self-contained rules**: Each rule module contains everything related to that rule
- **No breaking changes**: All existing code continues to work through re-exports
- **Easier maintenance**: Modifying a rule only requires touching one file
- **Clean separation**: Follows single responsibility principle

## Technical Details
- TOML parsing continues to work seamlessly
- All MD003 tests pass (25/25)
- All config tests pass (11/11)
- Build succeeds with no warnings
- CLI functionality verified with test files

## Future Work
This pattern can now be applied systematically to all remaining rule configurations,
eventually allowing each rule to be fully self-contained while maintaining
backward compatibility and unified configuration management.
